df-daemon: rewrite the cleanup code in initializer
the previous code on checking last access time is tricky and not
easy to read, instead we should make it more clearly that it is a
check on Stat_t.

Also, the df-daemon should deny non localhost access explicitly,
if this is reverse proxy for itself, then it is definitely a dead
loop.