options: add options to replace global commandline
It is generally a bad idea to use global variables, it makes it is
harder to write unit tests, and harder to track where they are used.
This is an effort to make the first move, and construct an options
so it can be pass down to the underlying system.