Add trust hosts params for dfdaemon
As dfdaemon be used to http_proxy mode, all the requests with host
are not local ip of dfdaemon then client will got a 403 error code.

Add a trust-hosts command line params and determine the request wether
be trusted or not, if it's trusted we directly forward them.

Related to issue:
https://github.com/dragonflyoss/Dragonfly/issues/385

Signed-off-by: godliness <machaojms@gmail.com>