feature: add transparent proxy to extend dfdaemon's proxy mode
Add a transparent proxy for dfdaemon. HTTP GET requests can be
proxied with dfget if it matches the proxy rule. Other methods,
including CONNECT for HTTPS requests will be proxied directly.

To handle CONNECT request properly, we can't use the default
http mux. So I created a new http server dedicated for proxy,
the port can be configured with '--proxyPort' cli flag, with a
default value of 65002. Proxy rules can be configured in
/etc/dragonfly/dfdaemon.yml

Signed-off-by: inoc603 <inoc603@gmail.com>